+++ ui/helper.c
@@ -30,7 +30,7 @@
 	#define ARRAY_SIZE(arr) (sizeof(arr)/sizeof((arr)[0]))
 #endif
 
-void UI_GenerateChannelString(char *pString, const uint8_t Channel)
+void UI_GenerateChannelString(char *pString, const uint16_t Channel)
 {
 	unsigned int i;
 
@@ -47,7 +47,7 @@ void UI_GenerateChannelString(char *pString, const uint8_t Channel)
 		pString[i + 3] = (gInputBox[i] == 10) ? '-' : gInputBox[i] + '0';
 }
 
-void UI_GenerateChannelStringEx(char *pString, const bool bShowPrefix, const uint8_t ChannelNumber)
+void UI_GenerateChannelStringEx(char *pString, const bool bShowPrefix, const uint16_t ChannelNumber)
 {
 	if (gInputBoxIndex > 0)
 	{
